requirements.txt
firebase-admin
django-firebase-orm
grpcio
djangorestframework